refactor: Moved dropdowns to their own subdirectory
